Improve Helm chart and jwt secret access logic (#3357)
This PR:

1. Improves the Helm chart configuration for easier and more reliable
deployment
2. Adds support for metrics-server integration with proper configuration
3. Enhances persistence management with better PVC/PV handling
4. Updates resource request configurations for better performance
5. Simplifies the deployment process with improved documentation and
defaults

---------
