+++ core/amber/src/main/scala/edu/uci/ics/texera/web/service/ResultExportService.scala
@@ -24,7 +24,6 @@ import java.nio.charset.StandardCharsets
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 import java.time.temporal.ChronoUnit
-import java.util.concurrent.{Executors, ThreadPoolExecutor}
 import java.util.zip.{ZipEntry, ZipOutputStream}
 import scala.collection.mutable
 import scala.util.Using
