+++ core/computing-unit-managing-service/src/main/scala/edu/uci/ics/texera/service/KubernetesConfig.scala
@@ -2,8 +2,6 @@ package edu.uci.ics.texera.service
 
 import com.typesafe.config.{Config, ConfigFactory}
 
-import scala.jdk.CollectionConverters.CollectionHasAsScala
-
 object KubernetesConfig {
 
   private val conf: Config = ConfigFactory.parseResources("kubernetes.conf").resolve()
