+++ deployment/k8s/texera-helmchart/Chart.yaml
@@ -41,3 +41,8 @@ dependencies:
   - name: lakefs
     version: 1.4.5
     repository: https://charts.lakefs.io
+
+  - name: metrics-server
+    version: 3.12.2
+    repository: https://kubernetes-sigs.github.io/metrics-server/
+    condition: metrics-server.enabled
