+++ deployment/k8s/texera-helmchart/templates/external-names.yaml
@@ -1,20 +1,51 @@
+{{/*
+Define a helper template for creating ExternalName services.
+This template takes three parameters:
+- name: The name of the service to create
+- namespace: The namespace where the service should be created
+- externalName: The fully qualified domain name to redirect to
+*/}}
+{{- define "external-name-service" -}}
 apiVersion: v1
 kind: Service
 metadata:
-  name: {{ .Values.fileService.name }}-svc
-  namespace: {{ .Values.workflowComputingUnitPool.namespace }}
+  name: {{ .name }}
+  namespace: {{ .namespace }}
 spec:
   type: ExternalName
-  externalName: {{ .Values.fileService.name }}-svc.{{ .Release.Namespace }}.svc.cluster.local
+  externalName: {{ .externalName }}
+{{- end }}
+
+{{/* Define namespace variables for better readability */}}
+{{- $namespace := .Release.Namespace }}
+{{- $workflowNamespace := .Values.workflowComputingUnitPool.namespace }}
+
+{{/* 
+Create ExternalName services in the workflow namespace to allow compute units
+to access services in the main namespace using the same service names.
+*/}}
+
+{{/* File service ExternalName */}}
+{{- include "external-name-service" (dict 
+  "name" (printf "%s-svc" .Values.fileService.name)
+  "namespace" $workflowNamespace
+  "externalName" (printf "%s-svc.%s.svc.cluster.local" .Values.fileService.name $namespace)
+) | nindent 0 }}
 
 ---
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ .Release.Name }}-postgresql
-  namespace: {{ .Values.workflowComputingUnitPool.namespace }}
-spec:
-  type: ExternalName
-  externalName: {{ .Release.Name }}-postgresql.{{ .Release.Namespace }}.svc.cluster.local
+{{/* PostgreSQL ExternalName */}}
+{{- include "external-name-service" (dict 
+  "name" (printf "%s-postgresql" .Release.Name)
+  "namespace" $workflowNamespace
+  "externalName" (printf "%s-postgresql.%s.svc.cluster.local" .Release.Name $namespace)
+) | nindent 0 }}
+
+---
+{{/* Webserver ExternalName */}}
+{{- include "external-name-service" (dict 
+  "name" (printf "%s-svc" .Values.webserver.name)
+  "namespace" $workflowNamespace
+  "externalName" (printf "%s-svc.%s.svc.cluster.local" .Values.webserver.name $namespace)
+) | nindent 0 }}
